+++ src/domManager.js
@@ -1,0 +1,19 @@
+import { saveTasksToLocalStorage } from './localStorageManager';
+
+function renderTasks() {
+  // Rest of the renderTasks code
+}
+
+function editPost(li, task) {
+  // Rest of the editPost code
+}
+
+function addNewTask(description) {
+  // Rest of the addNewTask code
+}
+
+function removeCompletedTasks() {
+  // Rest of the removeCompletedTasks code
+}
+
+export { renderTasks, editPost, addNewTask, removeCompletedTasks };
