--- src/localStorageManager.js
@@ -1,1 +1,0 @@
-
